# bot/src/main/resources/openapi/scrapper-api.yml
openapi: 3.1.0
info:
  title: Scrapper API
  version: 1.0.0
  contact:
    name: Alexander Biryukov
    url: https://github.com
paths:
  /tg-chat/{id}:
    post:
      summary: Зарегистрировать чат
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Чат зарегистрирован
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Чат уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    delete:
      summary: Удалить чат
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Чат успешно удалён
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Чат не существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /tg-chat/{id}/notification-mode:
    put:
      operationId: updateNotificationMode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationModeRequest'
      responses:
        '200':
          description: Режим обновлён
        '404':
          description: Чат не найден
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        description:
          type: string
        code:
          type: string
        exceptionName:
          type: string
        exceptionMessage:
          type: string
        stacktrace:
          type: array
          items:
            type: string
...
    NotificationModeRequest:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - INSTANT
            - DIGEST
        digestHour:
          type: integer
          minimum: 0
          maximum: 23